# use is_file() instead of file_exists() which is up to 100x faster