--- a/redaxo/src/addons/backup/lib/backup.php
+++ b/redaxo/src/addons/backup/lib/backup.php
@@ -484,7 +484,7 @@ class rex_backup
 
     private static function importScript($filename, $importType, $eventType)
     {
-        if (file_exists($filename)) {
+        if (is_file($filename)) {
             require $filename;
         }
     }