--- a/redaxo/src/addons/backup/pages/export.php
+++ b/redaxo/src/addons/backup/pages/export.php
@@ -54,9 +54,9 @@ if ($export && !$csrfToken->isValid()) {
         $ext = 'sql' == $exporttype ? '.sql' : '.tar.gz';
         $export_path = rex_backup::getDir() . '/';
 
-        if (file_exists($export_path . $filename . $ext)) {
+        if (is_file($export_path . $filename . $ext)) {
             $i = 1;
-            while (file_exists($export_path . $filename . '_' . $i . $ext)) {
+            while (is_file($export_path . $filename . '_' . $i . $ext)) {
                 ++$i;
             }
             $filename = $filename . '_' . $i;