--- a/redaxo/src/addons/install/lib/package/package_download.php
+++ b/redaxo/src/addons/install/lib/package/package_download.php
@@ -84,6 +84,6 @@ abstract class rex_install_package_download
             return $success;
         }
 
-        return file_exists("phar://$file/" . $this->addonkey);
+        return is_file("phar://$file/" . $this->addonkey);
     }
 }