--- a/redaxo/src/addons/install/lib/package/package_update.php
+++ b/redaxo/src/addons/install/lib/package/package_update.php
@@ -44,7 +44,7 @@ class rex_install_package_update extends rex_install_package_download
 
         // ---- check package.yml
         $packageFile = $temppath . rex_package::FILE_PACKAGE;
-        if (!file_exists($packageFile)) {
+        if (!is_file($packageFile)) {
             return rex_i18n::msg('package_missing_yml_file');
         }
         try {
@@ -58,7 +58,7 @@ class rex_install_package_update extends rex_install_package_download
         }
 
         // ---- include update.php
-        if ($this->addon->isInstalled() && file_exists($temppath . rex_package::FILE_UPDATE)) {
+        if ($this->addon->isInstalled() && is_file($temppath . rex_package::FILE_UPDATE)) {
             try {
                 $this->addon->includeFile('../.new.' . $this->addonkey . '/' . rex_package::FILE_UPDATE);
             } catch (rex_functional_exception $e) {