--- a/redaxo/src/addons/mediapool/pages/media.list.php
+++ b/redaxo/src/addons/mediapool/pages/media.list.php
@@ -297,7 +297,7 @@ $panel = '
                     }
 
                     // wenn datei fehlt
-                    if (!file_exists(rex_path::media($file_name))) {
+                    if (!is_file(rex_path::media($file_name))) {
                         $thumbnail = '<i class="rex-mime rex-mime-error" title="' . rex_i18n::msg('pool_file_does_not_exist') . '"></i><span class="sr-only">' . $file_name . '</span>';
                     } else {
                         $file_ext = substr(strrchr($file_name, '.'), 1);