--- a/redaxo/src/addons/phpmailer/install.php
+++ b/redaxo/src/addons/phpmailer/install.php
@@ -19,6 +19,6 @@ if ($addon->hasConfig('log')) {
 
 $oldBackUpFolder = rex_path::addonData('phpmailer', 'mail_backup');
 $logFolder = rex_path::addonData('phpmailer', 'mail_log');
-if (is_file($oldBackUpFolder) && !is_file($logFolder)) {
+if (is_dir($oldBackUpFolder) && !is_dir($logFolder)) {
     rename($oldBackUpFolder, $logFolder);
 }