--- a/redaxo/src/addons/structure/lib/structure_element.php
+++ b/redaxo/src/addons/structure/lib/structure_element.php
@@ -130,7 +130,7 @@ abstract class rex_structure_element
 
             $startId = rex_article::getSiteStartArticleId();
             $file = rex_path::addonCache('structure', $startId . '.1.article');
-            if (!rex::isBackend() && file_exists($file)) {
+            if (!rex::isBackend() && is_file($file)) {
                 // da getClassVars() eine statische Methode ist, können wir hier nicht mit $this->getId() arbeiten!
                 $genVars = rex_file::getCache($file);
                 unset($genVars['last_update_stamp']);