inspiriert von https://bugs.php.net/bug.php?id=78285

benchmark:
```php
<?php
$t0 = hrtime(true);
for ($i = 0; $i < 1000; $i++) {
    is_file(__FILE__);
}
$t1 = hrtime(true);
for ($i = 0; $i < 1000; $i++) {
    file_exists(__FILE__);
}
$t2 = hrtime(true);
printf("is_file      took %f ms\n", ($t1 - $t0)/10e6);
printf("file_exists took %f ms\n", ($t2 - $t1)/10e6);
```

ergibt auf meinem laptop
```sh
$ php perf.php
is_file      took 0.058450 ms
file_exists took 5.852210 ms

$ php perf.php
is_file      took 0.047010 ms
file_exists took 6.005810 ms

$ php -v
PHP 7.3.0 (cli) (built: Dec  6 2018 02:17:00) ( ZTS MSVC15 (Visual C++ 2017) x86
 )
Copyright (c) 1997-2018 The PHP Group
Zend Engine v3.3.0-dev, Copyright (c) 1998-2018 Zend Technologies
    with blackfire v1.24.1~win-x32-zts73, https://blackfire.io, by Blackfire
```

related https://github.com/symfony/symfony/issues/36493