@gharlan 2020-04-19
Verhalten die sich bei Symlinks gleich? (Gibt ja auch noch `is_link`)