@staabm 2020-04-19
Wo supporten/erwarten wir symlinks?