@gharlan 2020-04-19
> Wo supporten/erwarten wir symlinks?

YDeploy nutzt ja an einigen Stellen Symlinks (allerdings mehr für Ordner als für Dateien).

Ich glaube aber, dass sich da is_file/is_dir genauso verhalten wie file_exists, bin mir bloß eben nicht ganz sicher.