@staabm 2020-04-19
@bloep @gharlan könnt ihr den benchmark auch mal lokal testen?

> Es gibt noch mehr Stellen, wo es um Ordner geht. Man müsste also alle Stellen nochmal dahingehend durchgehen.

Guter Punkt - Sollte Jetzt passen.

> YDeploy nutzt ja an einigen Stellen Symlinks (allerdings mehr für Ordner als für Dateien).

Vllt  kannste das Mal kurz testen?