@bloep 2020-04-19
> könnt ihr den benchmark auch mal lokal testen?

```
$ php benchmark.php
is_file      took 0.172144 ms
file_exists took 0.283124 ms

$ php benchmark.php
is_file      took 0.020407 ms
file_exists took 0.287731 ms

$ php -v
PHP 7.4.1 (cli) (built: Jan  8 2020 20:29:32) ( NTS )
Copyright (c) The PHP Group
Zend Engine v3.4.0, Copyright (c) Zend Technologies
```
(aufm Macbook Air)