@bloep 2020-04-19
> Vllt hat auch jmd blackfire zur hand und kann den change messen

https://blackfire.io/profiles/compare/5fe27c63-e5c5-483c-a938-4cc2e335189d...e80ad6df-e9f2-42d5-9f72-af0473ddc433/graph
![image](https://user-images.githubusercontent.com/21292755/79691658-36bcc180-8261-11ea-82fa-412eae9e2480.png)