@staabm 2020-04-25
@gharlan lt. den kommentarem im oben verlinkten symfony issue scheint es so zu sein als würde die hier gemachte optimierung nur dort gut zu funktionieren, wo keine links im spiel sind (wie du urspr. vermutet hast).

d.h. müssen wir vermutlich alles was mit rex_path konfigurierbar ist, belassen wir vorher?

kannst du das im testing mit euren yakamara pfad schema bestätigten?