@gharlan 2020-04-25
> kannst du das im testing mit euren yakamara pfad schema bestätigten?

Die Symlinks entstehen ja erst beim Deployen, nicht lokal. Deswegen kann ich das nicht so ganz leicht testen. 
Wobei man einfach mal allgemein testen sollte, ob is_file für einen Symlink zu einer Datei auch true liefert, und is_dir für einen Symlink zu einem Ordner.